requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
airoframework==1.0.3
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
click==8.2.2
dotenv==0.9.9
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
yarl==1.20.1
